+++ packages/cli/src/colors/scale.ts
